+++ src/app/shells/main-window/components/settings-navigation-view/components/interface-settings/theme-primary-color/theme-primary-color.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ThemePrimaryColor } from './theme-primary-color';
+
+describe('ThemePrimaryColor', () => {
+  let component: ThemePrimaryColor;
+  let fixture: ComponentFixture<ThemePrimaryColor>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [ThemePrimaryColor]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(ThemePrimaryColor);
+    component = fixture.componentInstance;
+    await fixture.whenStable();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});