+++ release-notes.md
@@ -1,3 +1,7 @@
-# Fixes
+## Features
+
+* Application primary color selection.
+
+## Fixes
 
 * Improved icons preloading.